+++ src/Code/serverless/Client/Connector.js
@@ -1,11 +1,14 @@
 const uuidv4 = require("uuid/v4");
 const { ExternalZenatonError } = require("../../../Errors");
 
+const ZENATON_CONNECTOR_ID_PATTERN_LENGTH = 5;
+
 const Connector = class Connector {
   constructor(service, serviceId, processor) {
     this._checkString(service, "First", "connector's name");
     if (service !== "http") {
       this._checkString(serviceId, "Second", "connector's id");
+      this._checkServiceIdPattern(serviceId);
     }
 
     this._service = service;
@@ -71,6 +74,15 @@ const Connector = class Connector {
     if (typeof val !== "string" || val.length > 128) {
       throw new ExternalZenatonError(
         `${position} parameter of "${this._service}.${method}" (${type}) must have less than 128 characters"`,
+      );
+    }
+  }
+
+  // At this point we know that serviceId is a string because of ☝️
+  _checkServiceIdPattern(serviceId) {
+    if (serviceId.split("-").length !== ZENATON_CONNECTOR_ID_PATTERN_LENGTH) {
+      throw new ExternalZenatonError(
+        `The current given connectorId "${serviceId}" is not corresponding to a Zenaton connector Id, you should check at https://zenaton.com/documentation/node/api-connectors#use-connectors`,
       );
     }
   }